readonly mode fix

+++ src/pages/api/flow/[id].ts
@@ -24,6 +24,7 @@ export default async (req: NextApiRequest, res: NextApiResponse) => {
       logger.error("Flow ID not provided.");
       return res.status(400).json({ error: "Flow ID not provided" });
     }
+
     const flow = await prisma.flow.findUnique({
       where: { id: id as string },
       select: {
@@ -51,32 +52,40 @@ export default async (req: NextApiRequest, res: NextApiResponse) => {
     }
 
     logger.info("Flow retrieved successfully.");
+
     const token = await getToken({ req, secret: env.NEXTAUTH_SECRET });
     const { access_token } = await getOBOToken(token!);
     if (!access_token) {
-      res.status(403).json({ error: "Forbidden" });
-      return;
+      return res.status(403).json({ error: "Forbidden" });
     }
+
     const stardog = getStardogInstance({ token: access_token });
+
     switch (req.method) {
       case "GET":
         return res.status(200).json(flow);
 
       case "DELETE":
+        if (flow.userId !== user.id) {
+          logger.error("User not authorized to delete this flow.");
+          return res.status(403).json({ error: "Forbidden" });
+        }
+
         if (!flow.isDraft) {
           await stardog.deleteGraph(flow.name);
         }
 
-        const deleteFlow = await prisma.flow.delete({
+        const deletedFlow = await prisma.flow.delete({
           where: {
             id: id as string,
           },
         });
-        return res.status(200).json(deleteFlow);
+
+        logger.info("Flow deleted successfully.");
+        return res.status(200).json(deletedFlow);
 
       default:
         logger.error("Method not allowed.");
-
         return res.status(405).json({ error: "Method not allowed" });
     }
   } catch (error) {
